some error types are changed, public routes are  added
